fix: favicon, title, error handling, search on enter, constants rename
